+++ tools/build.R
@@ -9,7 +9,12 @@ devtools::load_all(".")
 # regenerate documentation
 dir.create("./docs")
 devtools::document()
-Rd2md::ReferenceManual(".", outdir = "./docs", verbose=FALSE)    
+Rd2md::ReferenceManual(".", outdir = "./docs", verbose=FALSE)
+api.doc.file <- "./docs/api.md"
+file.rename("./docs/Reference_Manual_..md", api.doc.file)
+api.doc <- readLines(api.doc.file)
+api.doc <- gsub(pattern = "```", replace = "\n```", x = api.doc)
+writeLines(api.doc, con=api.doc.file)
 
 # format code
 format.config <- list(recursive = TRUE, indent = 4, arrow = TRUE, brace.newline = FALSE, blank = TRUE)